Enable Redis Sessions
Added predis/predis package via composer

The Redis session requires a Redis instance. More information about the mxabierto/redis instance can be found in https://github.com/mxabierto/participa/issues/204

To test the Redis session it is necessary to do 'composer install' and change the .env variable 'SESSION_DRIVER' to 'redis'. After that, attempt to Login, Comment or perform an action on a document or comment and logout